Refactored DNS code into the dns server
